Unit tests, small fixes
