+++ client/sync.js
@@ -1,6 +1,7 @@
 'use strict'
 
 const initializer = require('./init')
+const bookmarkUtil = require('./bookmarkUtil')
 const RequestUtil = require('./requestUtil')
 const recordUtil = require('./recordUtil')
 const messages = require('./constants/messages')
@@ -96,75 +97,6 @@ const startSync = (requester) => {
     return jsRecords
   }
 
-  const getBaseBookmarksOrder = (deviceId, platform) => {
-    let platformPrefix = `1.`
-    if (platform === `ios` || platform === `android`) {
-      platformPrefix = `2.`
-    }
-    platformPrefix += deviceId + `.`
-    return platformPrefix
-  }
-
-  const getBookmarkOrder = (prevOrder, nextOrder) => {
-    let prevOrderSplit = prevOrder.split(`.`)
-    let nextOrderSplit = nextOrder.split(`.`)
-    if (prevOrderSplit.length === 1 && nextOrderSplit.length === 1) {
-      throw new Error(`Invalid previous and next orders: ${prevOrderSplit} and ${nextOrderSplit}`)
-    }
-    let order = ``
-    if (nextOrderSplit.length === 1) {
-    // Next order is an empty string
-      if (prevOrderSplit.length > 2) {
-        for (var i = 0; i < prevOrderSplit.length - 1; i++) {
-          order += prevOrderSplit[i] + `.`
-        }
-        order += (parseInt(prevOrderSplit[prevOrderSplit.length - 1]) + 1)
-      }
-    } else if (prevOrderSplit.length === 1) {
-      if (nextOrderSplit.length > 2) {
-        for (i = 0; i < nextOrderSplit.length - 1; i++) {
-          order += nextOrderSplit[i] + `.`
-        }
-        let lastNumber = parseInt(nextOrderSplit[nextOrderSplit.length - 1])
-        if (lastNumber === 1) {
-          order += `0.1`
-        } else {
-          order += (parseInt(nextOrderSplit[nextOrderSplit.length - 1]) - 1)
-        }
-      }
-    } else {
-      if (prevOrderSplit.length > 2 && nextOrderSplit.length > 2) {
-        for (i = 0; i < prevOrderSplit.length - 1; i++) {
-          order += prevOrderSplit[i] + `.`
-        }
-        if (prevOrderSplit.length === nextOrderSplit.length) {
-          let lastNumberPrev = parseInt(prevOrderSplit[prevOrderSplit.length - 1])
-          let lastNumberNext = parseInt(nextOrderSplit[nextOrderSplit.length - 1])
-          if (lastNumberNext - lastNumberPrev > 1) {
-            order += (lastNumberPrev + 1)
-          } else {
-            order += lastNumberPrev + `.1`
-          }
-        } else if (prevOrderSplit.length < nextOrderSplit.length) {
-          order += prevOrderSplit[prevOrderSplit.length - 1] + `.`
-          let currentIndex = prevOrderSplit.length
-          while (parseInt(nextOrderSplit[currentIndex]) === 0) {
-            order += nextOrderSplit[currentIndex] + `.`
-            currentIndex++
-          }
-          if (parseInt(nextOrderSplit[currentIndex]) === 1) {
-            order += `0.1`
-          } else if (parseInt(nextOrderSplit[currentIndex]) !== 0) {
-            order += (parseInt(nextOrderSplit[currentIndex]) - 1)
-          }
-        } else {
-          order += (parseInt(prevOrderSplit[prevOrderSplit.length - 1]) + 1)
-        }
-      }
-    }
-    return order
-  }
-
   ipc.on(messages.FETCH_SYNC_RECORDS, (e, categoryNames, startAt, limitResponse) => {
     logSync(`fetching ${categoryNames} records after ${startAt}`)
     categoryNames.forEach((category) => {
@@ -247,12 +179,12 @@ const startSync = (requester) => {
   })
   ipc.on(messages.GET_BOOKMARKS_BASE_ORDER, (e, deviceId, platform) => {
     logSync(`Getting bookmarks base order`)
-    ipc.send(messages.SAVE_BOOKMARKS_BASE_ORDER, getBaseBookmarksOrder(deviceId, platform))
+    ipc.send(messages.SAVE_BOOKMARKS_BASE_ORDER, bookmarkUtil.getBaseBookmarksOrder(deviceId, platform))
   })
   ipc.on(messages.GET_BOOKMARK_ORDER, (e, prevOrder, nextOrder) => {
     logSync(`Getting current bookmark order based on prev and next orders`)
 
-    ipc.send(messages.SAVE_BOOKMARK_ORDER, getBookmarkOrder(prevOrder, nextOrder), prevOrder, nextOrder)
+    ipc.send(messages.SAVE_BOOKMARK_ORDER, bookmarkUtil.getBookmarkOrder(prevOrder, nextOrder), prevOrder, nextOrder)
   })
   ipc.send(messages.SYNC_READY)
   logSync('success')
